change discount to percentage based

+++ src/main/java/models/Event.java
@@ -8,7 +8,7 @@ public class Event {
     private int foodCost;
     private int drinkCost;
     private int entertainmentCost;
-    private int eventCost;
+    private double eventCost;
 
     public void setGuestCount(int count) {
         this.guestCount = count;
@@ -88,8 +88,9 @@ public class Event {
         this.eventCost = this.foodCost + this.drinkCost + this.entertainmentCost;
     }
 
-    public int getEventCost() {
+    public double getEventCost() {
         return eventCost;
     }
+
 
 }
